mlserver/requirements.txt
sklearn
pandas
numpy
scipy
django-solo
django
djangorestframework
